add main files.
config file to read main configs like csv headers.
csvToSqlite is for saving in data base. You can save directly to sqlite But we chose to Do It like this :)
news.csv is for knowing that this crawler is working.
pagefailed and errorfile is for knowing which pages or which links failed to request.